pyproject.toml
[tool.poetry]
name = "compare_ai"
version = "0.0.1"
description = "A framework for comparing and using different AI models"
authors = ["santoshdahal2016 <webtutor2016@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typing-extensions = "^4.0.0"
pydantic = "^2.0.0"

# Optional dependencies for different providers
openai = { version = "^1.35.8", optional = true }
python-dotenv = "^1.0.1"
datasets = "^3.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
openai = ["openai"]
all = ["openai"]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
